Refactoring I/O (10) : Working through bugs (2), pretty confident I've smashed most of them.
Also added DefaultAnimCharMap to create a standard point where I can alter/extend the default map of char -> Index in Lexical-based Animation things

+++ src/main/kotlin/spirite/core/file/save/SifImgdWriter.kt
@@ -56,7 +56,7 @@ object SifImgdWriter {
                             }
                         }
                     }
-                    out.writeShort(medium.data.rawImgData.size)
+                    out.writeInt(medium.data.rawImgData.size)
                     out.writeShort(medium.data.offsetX.i)
                     out.writeShort(medium.data.offsetY.i)
                     out.write(medium.data.rawImgData)
